refactor: update verifier metdata as per response mdoe
